+++ Models/DomainModels/SubRegion.cs
@@ -11,7 +11,7 @@ namespace Cares.Models.DomainModels
         /// <summary>
         /// Sub Region ID
         /// </summary>
-        public int SubRegionId { get; set; }
+        public short SubRegionId { get; set; }
         /// <summary>
         /// Sub Region Code
         /// </summary>
@@ -31,7 +31,7 @@ namespace Cares.Models.DomainModels
         /// <summary>
         /// Region ID
         /// </summary>
-        public int RegionId { get; set; }
+        public short RegionId { get; set; }
         /// <summary>
         /// Row Version
         /// </summary>
